+++ spring-cloud-tencent-rpc-enhancement/src/test/java/com/tencent/cloud/rpc/enhancement/config/RpcEnhancementAutoConfigurationTest.java
@@ -17,11 +17,14 @@
 
 package com.tencent.cloud.rpc.enhancement.config;
 
+import java.io.IOException;
+import java.net.URI;
+
 import com.tencent.cloud.polaris.context.config.PolarisContextAutoConfiguration;
 import com.tencent.cloud.rpc.enhancement.feign.EnhancedFeignBeanPostProcessor;
+import com.tencent.cloud.rpc.enhancement.feign.EnhancedFeignPluginRunner;
 import com.tencent.cloud.rpc.enhancement.feign.plugin.reporter.ExceptionPolarisReporter;
 import com.tencent.cloud.rpc.enhancement.feign.plugin.reporter.SuccessPolarisReporter;
-import com.tencent.cloud.rpc.enhancement.resttemplate.EnhancedRestTemplateModifier;
 import com.tencent.cloud.rpc.enhancement.resttemplate.EnhancedRestTemplateReporter;
 import com.tencent.polaris.api.core.ConsumerAPI;
 import org.junit.Test;
@@ -29,6 +32,11 @@ import org.junit.Test;
 import org.springframework.boot.autoconfigure.AutoConfigurations;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.context.runner.WebApplicationContextRunner;
+import org.springframework.cloud.client.ServiceInstance;
+import org.springframework.cloud.client.loadbalancer.LoadBalanced;
+import org.springframework.cloud.client.loadbalancer.LoadBalancerAutoConfiguration;
+import org.springframework.cloud.client.loadbalancer.LoadBalancerClient;
+import org.springframework.cloud.client.loadbalancer.LoadBalancerRequest;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
@@ -46,18 +54,22 @@ public class RpcEnhancementAutoConfigurationTest {
 			.withConfiguration(AutoConfigurations.of(
 					PolarisContextAutoConfiguration.class,
 					RpcEnhancementAutoConfiguration.class,
-					PolarisRestTemplateAutoConfigurationTester.class))
+					PolarisRestTemplateAutoConfigurationTester.class,
+					LoadBalancerAutoConfiguration.class))
 			.withPropertyValues("spring.cloud.polaris.circuitbreaker.enabled=true");
 
 	@Test
 	public void testDefaultInitialization() {
 		this.contextRunner.run(context -> {
 			assertThat(context).hasSingleBean(ConsumerAPI.class);
+			assertThat(context).hasSingleBean(EnhancedFeignPluginRunner.class);
 			assertThat(context).hasSingleBean(EnhancedFeignBeanPostProcessor.class);
 			assertThat(context).hasSingleBean(SuccessPolarisReporter.class);
 			assertThat(context).hasSingleBean(ExceptionPolarisReporter.class);
-			assertThat(context).hasSingleBean(EnhancedRestTemplateModifier.class);
 			assertThat(context).hasSingleBean(EnhancedRestTemplateReporter.class);
+			assertThat(context).hasSingleBean(RestTemplate.class);
+			RestTemplate restTemplate = context.getBean(RestTemplate.class);
+			assertThat(restTemplate.getErrorHandler() instanceof EnhancedRestTemplateReporter).isTrue();
 		});
 	}
 
@@ -66,8 +78,34 @@ public class RpcEnhancementAutoConfigurationTest {
 	static class PolarisRestTemplateAutoConfigurationTester {
 
 		@Bean
+		@LoadBalanced
 		RestTemplate restTemplate() {
 			return new RestTemplate();
+		}
+
+		@Bean
+		LoadBalancerClient loadBalancerClient() {
+			return new LoadBalancerClient() {
+				@Override
+				public ServiceInstance choose(String serviceId) {
+					return null;
+				}
+
+				@Override
+				public <T> T execute(String serviceId, LoadBalancerRequest<T> request) throws IOException {
+					return null;
+				}
+
+				@Override
+				public <T> T execute(String serviceId, ServiceInstance serviceInstance, LoadBalancerRequest<T> request) throws IOException {
+					return null;
+				}
+
+				@Override
+				public URI reconstructURI(ServiceInstance instance, URI original) {
+					return null;
+				}
+			};
 		}
 	}
 }